Rename: style 폴더 아래 home 관련 모듈 css style/home 아래로 이동
